refactor and introduce composable into custom assets flow
